Move router tests
